+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.5.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Change Add, Edit and Delete to be buttons instead of links.
+  See https://github.com/plone/Products.CMFPlone/issues/655.
+  [pbauer]
 
 
 1.5.1 (2015-06-03)